requirements.txt
streamlit==1.31.0
scikit-learn==1.4.0
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.3.1
plotly==5.18.0
pandas==2.1.4
numpy==1.26.4
spacy==3.7.2
